+++ DBOnDemand/src/java/ch/cern/dod/db/entity/DODCommandStat.java
@@ -16,7 +16,7 @@ public class DODCommandStat {
     /**
      * Mean duration of the command executions
      */
-    private float meanDuration;
+    private int meanDuration;
 
     public String getCommandName() {
         return commandName;
@@ -34,11 +34,11 @@ public class DODCommandStat {
         this.count = count;
     }
 
-    public float getMeanDuration() {
+    public int getMeanDuration() {
         return meanDuration;
     }
 
-    public void setMeanDuration(float meanDuration) {
+    public void setMeanDuration(int meanDuration) {
         this.meanDuration = meanDuration;
     }
 }
